Get thread data action
